backend/requirements.txt
fastapi
uvicorn
google-genai
google-generativeai
langgraph
python-dotenv
livekit-api
livekit-agents[images]
livekit-plugins-google
qdrant-client
pymupdf
hypothesis
python-multipart
python-docx
python-pptx
openpyxl
xlrd
striprtf
beautifulsoup4
asteval
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
hvac
